types check support typeparam
