+++ types_go117.go
@@ -1,0 +1,34 @@
+//go:build !go1.18
+// +build !go1.18
+
+package igop
+
+import (
+	"go/ast"
+	"go/types"
+)
+
+const (
+	EnabledTypeParam = false
+)
+
+func IsTypeParam(t types.Type) bool {
+	return false
+}
+
+func (sp *sourcePackage) Load() (err error) {
+	if sp.Info == nil {
+		sp.Info = &types.Info{
+			Types:      make(map[ast.Expr]types.TypeAndValue),
+			Defs:       make(map[*ast.Ident]types.Object),
+			Uses:       make(map[*ast.Ident]types.Object),
+			Implicits:  make(map[ast.Node]types.Object),
+			Scopes:     make(map[ast.Node]*types.Scope),
+			Selections: make(map[*ast.SelectorExpr]*types.Selection),
+		}
+		if err := types.NewChecker(sp.Context.conf, sp.Context.FileSet, sp.Package, sp.Info).Files(sp.Files); err != nil {
+			return err
+		}
+	}
+	return
+}
