+++ types_go118.go
@@ -1,0 +1,43 @@
+//go:build go1.18
+// +build go1.18
+
+package igop
+
+import (
+	"go/ast"
+	"go/types"
+)
+
+const (
+	EnabledTypeParam = true
+)
+
+func IsTypeParam(t types.Type) bool {
+	switch t := t.(type) {
+	case *types.TypeParam:
+		return true
+	case *types.Named:
+		return t.TypeParams() != nil
+	case *types.Signature:
+		return t.RecvTypeParams() != nil || t.TypeParams() != nil
+	}
+	return false
+}
+
+func (sp *sourcePackage) Load() (err error) {
+	if sp.Info == nil {
+		sp.Info = &types.Info{
+			Types:      make(map[ast.Expr]types.TypeAndValue),
+			Defs:       make(map[*ast.Ident]types.Object),
+			Uses:       make(map[*ast.Ident]types.Object),
+			Implicits:  make(map[ast.Node]types.Object),
+			Scopes:     make(map[ast.Node]*types.Scope),
+			Selections: make(map[*ast.SelectorExpr]*types.Selection),
+			Instances:  make(map[*ast.Ident]types.Instance),
+		}
+		if err := types.NewChecker(sp.Context.conf, sp.Context.FileSet, sp.Package, sp.Info).Files(sp.Files); err != nil {
+			return err
+		}
+	}
+	return
+}
